feat: enable DOM picking for HTML iframes
